Initial progress on S3 storage tooling
This is a first draft of an S3 syncing script that takes the contents of
the storage root and copies them (or a subdir of them) to S3. Right now
it doesn't have fancy exclusion rules or other features, just copies the
whole dirs file-by-file.
Unfortunately, this is the simplest way to do an upload to S3, and it's
pretty slow for many files which are not being updated. To speed things
up, we should switch to checking the ETag on any existing objects as the
default behavior.
Progress on #2