Use threadpool to speedup repo-s3-sync ~100x
Almost all of the time spent doing the upload/download work in this
script was wasted on single-threaded synchronous processing. However,
every file can be treated independently, so it's totally safe to divvy
this up into a threadpool of 100 threads (tunable if necessary). To make
printing work again, requires a little bit of cleverness and a single
thread which owns stdout.