Update repo-s3-sync to allow `--delete` syncs
This is desirable for handling the web content, which will likely
involve adding and removing files much more often than the repos do.
Tracks files to delete, ensures that the newly uploaded content is
impervious to harm, and reuses the threadpool to achieve speedy
deletions. Add `--delete` to the manpage for this tool.